fix: post creation
