+++ internal/guest/handler/customer/get_me.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/ua-academy-projects/share-bite/internal/util/httpctx"
+	"github.com/ua-academy-projects/share-bite/internal/guest/util/httpctx"
 )
 
 func (h *handler) getMe(c *gin.Context) {
