+++ internal/guest/util/httpctx/httpctx.go
@@ -4,26 +4,29 @@ import (
 	"fmt"
 
 	"github.com/gin-gonic/gin"
+	"github.com/ua-academy-projects/share-bite/internal/middleware"
 )
 
-const (
-	userIDContextKey = "userId"
-)
+func Get[T any](c *gin.Context, key string) (T, error) {
+	var zero T
+
+	val, ok := c.Get(key)
+	if !ok {
+		return zero, fmt.Errorf("%s missing in context", key)
+	}
+
+	typedVal, ok := val.(T)
+	if !ok {
+		return zero, fmt.Errorf("%s invalid type in context", key)
+	}
+
+	return typedVal, nil
+}
 
 func GetUserID(c *gin.Context) (string, error) {
-	val, ok := c.Get(userIDContextKey)
-	if !ok {
-		return "", fmt.Errorf("userId missing in context")
-	}
+	return Get[string](c, middleware.CtxUserID)
+}
 
-	userID, ok := val.(string)
-	if !ok {
-		return "", fmt.Errorf("userId cast failed")
-	}
-
-	if userID == "" {
-		return "", fmt.Errorf("userId is empty")
-	}
-
-	return userID, nil
+func GetUserRole(c *gin.Context) (string, error) {
+	return Get[string](c, middleware.CtxUserRole)
 }
